Added custom properties to CSS, disabled body scrolling when popup active
The addition of some custom properties to the CSS file will now make basic customisation easier.

When the popup is active, the overflow of the body is set to hidden. This stops the irritation of the body being scrollable until popup is closed.

README.md has been updated to explain the custom properties in the CSS file