Mainly start of zoom on gallery component
We have added the ability to zoom in and out using the mouse wheel.
The image will remain centred, we have to add our 'drag' to tate full advantage of this feature.
Some debugging console logs removed we no longer need.
Some surplus debugging tags in the php file removed as well.

Som