Drag ability added to gallery
By holding down the left mouse button and moving with the mouse, we can now drag the image in the gallery window.
I need to test this with touch and there is some difficulty getting it to null the offsets when the image is cycled, it loads fine but when yuo drag the new image it jumps a bit.