Fixed bug with position shift when image rotated, began touch-drag controls
The image would shift on a drag action  when the previous image had been dragged. This is fixed now!

Touch will now perform all operations available to mouse but there is a small bug,
when using touch, there is no .movementX , .movementY. Therefore we have to make our own delta x & y using stored offsets.
On a second touch, these reset but that means the image jumps a little.. I need the previous offsets to persist, shouldn't be a major problem to fix!