Touch controls all work now plus some other teething issues sorted
Additional offsets were required to make things work with drag on touchscreen
undesired persistence of offsets etc caused problems when popup was activated for a second time onward.
Using the existing setInterval to monitor for it's close and reset proved effective enough.
Because of the animations, it's very unlikely the popup will be deactivated and re-activated inside 1000ms (1s), perfect timing would be needed to get the full period anyway!.