Added prompt displaying which image is currently displayed (eg: Image 1 of 4)
There was no way for the user to see how many images there were to scroll through, nor where they were in relation to them.
This was very awkward when only one image was used, the user would not be able to tell if the component was functioning correctly.
By adding this to the controls which fade in and out according to movement, this will not impede view of images.