Addition of loading "spinner"
This is my favourite change. I have used an external component for this, a simple animated spinner made with HTML and CSS.
The link for this is provided in the source and within the gallery readme.md.
The spinner is displayed while the file is loading. This makes the gallery much more user friendly on slower connections.
Before, it was easy tho think something was wrong when there was no response having clicked the arrow to change images etc.