Multiple popup capability added to popup component
This isn't strictly true. The component has been changed to allow content to be dynamically swapped out at runtime by loading an external html file as the content.

default content is placed in default.html and any other content in a different html file. These files are within the popup subdirectory.

The title is passed as the first parameter to the activatePopup function, the filename minus extension is an optional second parameter which defaults to 'default'.

The fetch API is then used to read the file and place it as the innerHTML of the pop up body.

This allows us to call the pop up with different content of choice. It is easy to add id's and classes to the html within the content files and as long as they are referenced AFTER the content is loaded, they will be available to allow for dynamic content within the pop up itself!